[#1512] Factor out Customer class methods
Also "includes" the class includes (no pun intended).
